extracted status update
